refactor hook for fetching and calculating portfolio positions and breakdown into hook + documentation comments
